+++ management/CHANGELOG.md
@@ -1,8 +1,13 @@
-# v0.64.0 (Unreleased)
+# v0.65.0 (Unreleased)
 
 * **Note** Remove redundant data models. [#501](https://github.com/patrickcping/pingone-go-sdk-v2/pull/501)
 * **Enhancement** Re-aligned the `FormFieldQrCode` model to the API contract for DaVinci Forms. [#501](https://github.com/patrickcping/pingone-go-sdk-v2/pull/501)
 * **Enhancement** Added support for `FormFieldPolling`, `FormFieldFIDO2`, `FormFieldSingleCheckbox` and `FormFieldAgreement` data models for DaVinci Forms. [#501](https://github.com/patrickcping/pingone-go-sdk-v2/pull/501)
+* **Enhancement** Added `filter` parameter to language translations API. [#511](https://github.com/patrickcping/pingone-go-sdk-v2/pull/511)
+
+# v0.64.0 (2026-02-26)
+
+* **Enhancement** Added `AlphabetSequenceRule`, `NumberSequenceRule`, `ShiftedNumberRowSequenceRule`, `QwertySequenceRule` to the `PasswordPolicy` data model.  Removed `BypassPolicy`, `CurrentPassword`, `NewPassword` from the `PasswordPolicy` data model per API contract. [#502](https://github.com/patrickcping/pingone-go-sdk-v2/pull/502)
 
 # v0.63.0 (2025-12-11)